Add option to ignore foreign key checks.
Adds an 'ignoreKeys = true' option to the default options which will disable foreign key constraints for the mysql dump. Foreign key constraints are re-enabled once the import of the dump is completed.

Implemented as dumps that include data may require that tables are restored in a particular order to satisfy foreign key constraints.