SET foreign_key_checks = 0;
I found and solved an issue.
I have a db ordersmanager with two tables: orders and suppliers.
orders has a column, supplierId, wich is a foreign key to the table suppliers. supplierId is indeed the suppliers' primary key.
Now, if I put a supplier with { supplierId: "AAA", name: "Bob" } and then a order { orderId: "OOO", supplierId: "AAA" } it works, but if I put first the order { orderId: "OOO", supplierId: "AAA" }  and then the supplier  { supplierId: "AAA", name: "Bob" }, it outputs an error, because when I put the order, there is no supplier with id AAA yet.
mysqldump does not take care of this, indeed, in my db .sql backup, it first insert the orders and after the suppliers, resulting in an error.
I have modified the module, now there is the option disableForeignKeyChecks, the default value is false but if it is setted to true then SET foreign_key_checks = 0; is added as the first line of the generated .sql backup file. I tested it and it works. I suggest you, if you want to merge this feature, to modify the readme.md too.