keep the table order of the table whitelist
When providing a table whitelist the order of that whitelist should be kept when dumping the tables.

This helps if dumping the database without table locks and having tables with foreign key references into other tables. By carefully defining the order of tables you can still achieve a consistent dump without violating the referential integrity by first dumping the table containing the foreign key, followed by the table which is the target of the reference.